* commit '4b4cc70bea6e54ef52abb6e61fee7f207a8d359b':
  PI-3182: java library cleanup: more fixes

+++ src/main/java/com/impinj/itemsense/client/coordinator/readerconfiguration/ChannelConfig.java
@@ -13,6 +13,6 @@ import lombok.NoArgsConstructor;
 public class ChannelConfig {
 
   private int channelIndex;
-  private int[] txFrequenciesInMhz;
+  private double[] txFrequenciesInMhz;
 
 }
